Moved away from log4j to take ownership of how logging is done here
This will give us some performance problems and logs will grow
endlessly, but these can be fixed. For now it is probably better
to just take complete ownership of the logging done here, since
it is a well defined part of output. This way we also won't be
messing with any clients of this library.