Resolves issue #300
The issue was causing errors when running more than one command. Buffer
reuse was causing a token to be detected and "skipped" before it was
actually consumed. This caused an early close of the stream.

This was primarily provoked by having many files of the same size.
Causing a very similar flow, so similar, that the EOF tokens would be
placed on the same offset of the buffer.